Sixteen commits of main land on 160 of the branch's own. 763 files
merged cleanly; of 172 conflict hunks, 153 were main reformatting code
the branch had renamed, so the branch's text stands there.

The rest were real: digression keeps main's TASTy frame resolver and
its display fields, stenography keeps the Designator rename and the
new functionSyntax/methodSyntax helpers, decorum takes main's rewrite
whole, and geodesy unions the imports. build.mill takes main's new
modules -- kotlin, asn1, x509, geodesy.angle -- expressed in the
branch's vocabulary (no scala in -Yimports, settings.sep) and keeps
the branch's murmuration module, which main never had, and its p8
compiler pin, which the opaque Array depends on.

Main's new code then had to learn the branch's vocabulary: asn1's
parser becomes honestly Mutable with explicit Tactic clauses, IArray
literals become frozen Arrays, JDK seams take Array.unsafeJvm, and
Optional flattening goes through vacuous's compact.

+++ lib/symbolism/src/core/symbolism.Multiplicable.scala
@@ -60,7 +60,7 @@ object Multiplicable:
     (multiplicand, multiplier) => multiplicand*multiplier
 
   given short: Short is Multiplicable by Short to Short = Multiplicable:
-    (multiplicand, multiplier) => (multiplicand*multiplier).toByte
+    (multiplicand, multiplier) => (multiplicand*multiplier).toShort
 
   given byte: Byte is Multiplicable by Byte to Byte = Multiplicable:
     (multiplicand, multiplier) => (multiplicand*multiplier).toByte
